move dialog functions from app.rs into new file app_dialogs.rs
